// arg-web-frontend/src/types/accounts-payable.types.ts
export interface CustomColumnType<T> extends ColumnType<T> {
  filterable?: boolean; // ✅ not `filter`, not `filtered`
}
// ...
export interface DataType {
  key: string;
  carrierId: string;
  carrierInvoice: string;
  orderShipDate: string;
  invType: number;
  orderNo: number;
  shipRef: number;
  invoiceAmount: string;
}